Add required dependencies for Next.js app
This commit adds the necessary dependencies for the Next.js application. With the new setup, tools required for the operations are included. This makes the application more autonomous by allowing it to track its dependencies and manage its operations.Add .idea to .gitignore

This commit adds JetBrains IDE's configuration directory (.idea) to .git